--- app/components/post/AlbumArtCard.module.css
@@ -1,35 +1,0 @@
-.albumImageContainer {
-  position: relative;
-  width: 300px;
-  height: 300px;
-  transition: all 0.1s;
-}
-
-.albumImageOverlay {
-  position: absolute;
-  width: 300px;
-  height: 300px;
-  background: linear-gradient(
-    105deg,
-    transparent 40%,
-    rgba(255, 255, 255, 0.8) 45%,
-    /* White with high opacity */ rgba(255, 255, 255, 0.6) 50%,
-    /* White with medium opacity */ transparent 54%
-  );
-  filter: brightness(1.1) opacity(0.8);
-  mix-blend-mode: color-dodge;
-  background-size: 150% 150%;
-  background-position: 100%;
-  transition: all 0.1s;
-}
-
-.albumImage {
-  cursor: pointer;
-  width: 100%;
-  height: 100%;
-  background-size: contain;
-  box-shadow:
-  /* Inner shadow for texture */
-    inset 0 0 5px rgba(0, 0, 0, 0.3),
-    /* Outer shadow for depth */ 0 4px 8px rgba(0, 0, 0, 0.2);
-}
