+++ app/components/post/PostMetadata.tsx
@@ -1,5 +1,5 @@
 import { usePathname } from "next/navigation";
-import styles from "./PostAlbumMetadata.module.css";
+import styles from "./PostMetadata.module.css";
 import { formatDate, getFormattedDuration, isAdminPage } from "../../modules/utils";
 import { DeleteButton } from "../@common/assets/DeleteButton";
 import { EditButton } from "../@common/assets/EditButton";
@@ -7,15 +7,12 @@ import Link from "next/link";
 import { LinkIcon } from "../@common/assets/LinkIcon";
 import { BlurImg } from "../@common/BlurImg";
 import { AlbumInfo } from "../../modules/types";
-// import { AlbumArtCard } from "./AlbumArtCard";
-// import { isBrowser } from "react-device-detect";
-// import { useEffect, useState } from "react";
 
-interface PostAlbumMetadataProps {
+interface PostMetadataProps {
   postData: AlbumInfo;
 }
 
-export const PostAlbumMetadata = ({ postData }: PostAlbumMetadataProps) => {
+export const PostMetadata = ({ postData }: PostMetadataProps) => {
   const {
     id,
     videos,
@@ -33,11 +30,6 @@ export const PostAlbumMetadata = ({ postData }: PostAlbumMetadataProps) => {
   const albumDuration = getFormattedDuration(duration);
   const pathName = usePathname();
   const hasVideo = videos[0]?.title.length > 0;
-  // const [browserView, setBrowserView] = useState(false);
-
-  // useEffect(() => {
-  //   isBrowser && setBrowserView(true);
-  // }, []);
 
   return (
     <header className={styles.container}>
@@ -47,16 +39,6 @@ export const PostAlbumMetadata = ({ postData }: PostAlbumMetadataProps) => {
           <BlurImg className={styles.albumImage} blurHash={blurHash} src={imgUrl} punch={1} />
         </a>
       </div>
-      {/* {browserView ? (
-        <AlbumArtCard link={link} imgUrl={imgUrl} />
-      ) : (
-        <div className={styles.albumImageContainer}>
-          <a href={link} target="_blank">
-            <BlurImg className={styles.albumImage} blurHash={blurHash} src={imgUrl} punch={1} />
-          </a>
-        </div>
-      )} */}
-
       <div className={styles.metadataContainer}>
         {/* 아티스트 정보 */}
         <label className={styles.metadataTitle}>아티스트</label>
