+++ app/components/post/comment/@common/ManageButtons.module.css
@@ -1,4 +1,4 @@
-.buttonContainer {
+.container {
   position: absolute;
   display: flex;
   flex-direction: column;
@@ -15,7 +15,7 @@
   align-items: center;
   width: 90px;
   height: 30px;
-  border: 1px solid var(--second-text-color);
+  border: 1px solid var(--text-color);
   transition: background-color 0.3s;
 }
 
@@ -23,7 +23,7 @@
   color: #000;
   font-weight: bold;
   border-radius: 0;
-  background-color: var(--second-text-color);
+  background-color: var(--text-color);
 }
 
 .dotButton {
