+++ app/components/post/comment/Comments.module.css
@@ -1,12 +1,5 @@
 .container {
   display: flex;
   flex-direction: column;
-  padding: 0 30px;
   box-sizing: border-box;
-}
-
-@media (max-width: 1050px) {
-  .container {
-    padding: 0;
-  }
 }
