+++ app/components/post/comment/like/LikeIcon.module.css
@@ -4,14 +4,7 @@
   column-gap: 5px;
 }
 
-.heartImage {
-  background-position: center;
-  background-repeat: no-repeat;
-  width: 20px;
-  height: 20px;
-}
-
-.heartSvgFile {
+.likeIcon {
   stroke-width: 12px;
   cursor: pointer;
   margin-top: 1px;
