+++ app/components/post/comment/reply/ReplyInput.module.css
@@ -2,6 +2,7 @@
   display: flex;
   flex-direction: column;
   column-gap: 15px;
+  padding: 0 0 0 60px;
 }
 
 .commentContainer {
@@ -13,8 +14,8 @@
 }
 
 .userImageWrapper {
-  min-width: 44px;
-  height: 44px;
+  min-width: 45px;
+  height: 45px;
   overflow: hidden;
   border-radius: 50%;
 }
@@ -33,7 +34,7 @@
   align-items: flex-end;
   width: 100%;
   background-color: var(--input-background-color);
-  border: 1px solid var(--second-text-color);
+  border: 1px solid var(--text-color);
   padding: 10px;
   box-sizing: border-box;
   row-gap: 10px;
@@ -45,7 +46,7 @@
 }
 
 .textarea {
-  border: 1px solid var(--second-text-color);
+  border: 1px solid var(--text-color);
   min-height: 100px;
   width: 100%;
   padding: 8px;
@@ -57,7 +58,7 @@
 
 .button {
   width: fit-content;
-  border: 1px solid var(--second-text-color);
+  border: 1px solid var(--text-color);
   border-color: #1b1f2326;
   appearance: none;
   border: 1px solid;
@@ -71,7 +72,7 @@
   user-select: none;
   vertical-align: middle;
   white-space: nowrap;
-  color: var(--second-text-color);
+  color: var(--text-color);
   background-color: var(--input-background-color);
 }
